Split off delayed tracking service into its own class and some other refactorings
