+++ src/main/resources/application-dev.yml
@@ -1,15 +1,6 @@
 ceres:
   downsample:
     granularities:
-      - width: 5s
-        ttl: 1h
-        partitionWidth: 1m
-      - width: 10s
-        ttl: 1h
-        partitionWidth: 1m
-      - width: 30s
-        ttl: 1h
-        partitionWidth: 1m
       - width: 1m
         ttl: 1h
         partitionWidth: 5m
